[SW-87] refactor: 백엔드 코드 정리
- 소셜 로그인
