+++ backend/src/main/java/com/done/swim/domain/user/service/UserService.java
@@ -5,7 +5,7 @@ import com.done.swim.domain.user.entity.User;
 import com.done.swim.domain.user.repository.UserRepository;
 import com.done.swim.global.exception.ErrorCode;
 import com.done.swim.global.exception.ResourceNotFoundException;
-import com.done.swim.global.jwt.JwtTokenProvider;
+import com.done.swim.oauth2.service.OAuth2TokenService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -16,7 +16,7 @@ import org.springframework.transaction.annotation.Transactional;
 public class UserService {
 
     private final UserRepository userRepository;
-    private final JwtTokenProvider jwtTokenProvider;
+    private final OAuth2TokenService oAuth2TokenService; // 토큰 관리 서비스 주입
 
     /**
      * 유저 정보 조회
@@ -30,5 +30,22 @@ public class UserService {
 
         return UserInfoResponseDto.from(user);
 
+    }
+
+    /**
+     * 회원탈퇴
+     */
+    @Transactional
+    public void deleteUserWithTokens(Long userId) {
+
+        // 유저 존재 여부 확인
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new IllegalArgumentException("해당 유저가 존재하지 않습니다."));
+
+        // Redis에서 리프레시 토큰 삭제
+        oAuth2TokenService.deleteRefreshToken(userId);
+
+        // 유저 정보 삭제
+        userRepository.delete(user);
     }
 }
