+++ backend/src/main/java/com/done/swim/global/config/SecurityConfig.java
@@ -4,9 +4,9 @@ package com.done.swim.global.config;
 import com.done.swim.global.jwt.JwtAuthenticationFilter;
 import com.done.swim.global.security.handler.CustomAccessDeniedHandler;
 import com.done.swim.global.security.handler.JwtAuthenticationEntryPoint;
-import com.done.swim.oauth2.CustomOAuth2UserService;
 import com.done.swim.oauth2.handler.OAuth2LoginFailureHandler;
 import com.done.swim.oauth2.handler.OAuth2LoginSuccessHandler;
+import com.done.swim.oauth2.service.CustomOAuth2UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
