+++ backend/src/main/java/com/done/swim/global/security/handler/CustomAccessDeniedHandler.java
@@ -2,6 +2,7 @@ package com.done.swim.global.security.handler;
 
 
 import com.done.swim.common.ApiResponse;
+import com.done.swim.global.exception.ErrorCode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
@@ -25,7 +26,7 @@ public class CustomAccessDeniedHandler implements AccessDeniedHandler {
         response.setStatus(HttpServletResponse.SC_FORBIDDEN);
 
 
-        ApiResponse<Void> errorResponse = ApiResponse.error("접근 권한이 없습니다.", "FORBIDDEN");
+        ApiResponse<Void> errorResponse = ApiResponse.error(ErrorCode.DEFAULT_FORBIDDEN.getMessage(), ErrorCode.DEFAULT_FORBIDDEN.getCode());
         response.getWriter().write(objectMapper.writeValueAsString(errorResponse));
     }
 }
