+++ backend/src/main/java/com/done/swim/oauth2/handler/OAuth2LoginFailureHandler.java
@@ -28,8 +28,6 @@ public class OAuth2LoginFailureHandler extends SimpleUrlAuthenticationFailureHan
 
         // 로그인 실패 시 리디렉트할 URL 설정
         getRedirectStrategy().sendRedirect(request, response, "/login");
-//        //  로그인 실패 시 리디렉트할 URL 설정
-//        response.sendRedirect("/loginFailure");
 
     }
 }
