+++ backend/src/main/java/com/done/swim/oauth2/provider/CustomOAuth2User.java
@@ -1,11 +1,12 @@
-package com.done.swim.oauth2;
+package com.done.swim.oauth2.provider;
 
 import com.done.swim.domain.user.entity.User;
-import java.util.Collection;
-import java.util.Map;
 import lombok.Getter;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.oauth2.core.user.OAuth2User;
+
+import java.util.Collection;
+import java.util.Map;
 
 @Getter
 public class CustomOAuth2User implements OAuth2User {
