+++ backend/src/main/java/com/done/swim/oauth2/provider/GithubUserInfo.java
@@ -31,7 +31,6 @@ public class GithubUserInfo implements OAuth2UserInfo {
 
     @Override
     public String getEmail() {
-//        return (String) attributes.get("email");
         return email;
     }
