+++ backend/src/main/java/com/done/swim/oauth2/service/OAuth2TokenService.java
@@ -1,6 +1,5 @@
-package com.done.swim.oauth2;
+package com.done.swim.oauth2.service;
 
-import com.done.swim.common.ApiResponse;
 import com.done.swim.domain.user.entity.User;
 import com.done.swim.domain.user.repository.UserRepository;
 import com.done.swim.global.exception.ErrorCode;
@@ -12,10 +11,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseCookie;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -37,22 +34,26 @@ public class OAuth2TokenService {
     @Value("${cookie.secure}")
     private boolean secure;
 
-    // 리프레시 토큰 저장(redis)
+    /**
+     * 리프레시 토큰 저장(redis)
+     */
     public void saveRefreshToken(Long userId, String refreshToken) {
         redisTemplate.opsForValue()
                 .set("REFRESH_TOKEN:" + userId, refreshToken, refreshTokenValidityInMilliseconds,
                         TimeUnit.MILLISECONDS);
     }
 
-    // 리프레시 토큰 조회
+    /**
+     * 리프레시 토큰 조회
+     */
     public String getRefreshToken(Long userId) {
         return redisTemplate.opsForValue().get("REFRESH_TOKEN:" + userId);
     }
 
-    // 로그아웃 처리 (리프레시 토큰 삭제 + SecurityContext 초기화 + 쿠키 삭제)
+    /**
+     * 로그아웃 처리 (리프레시 토큰 삭제 + SecurityContext 초기화 + 쿠키 삭제)
+     */
     public ResponseEntity<String> logout(User user, HttpServletResponse response) {
-
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
 
         // 레디스에서 리프레시 토큰 삭제
         redisTemplate.delete("REFRESH_TOKEN:" + user.getId());
@@ -74,7 +75,9 @@ public class OAuth2TokenService {
         return ResponseEntity.ok("로그아웃 성공");
     }
 
-    // 액세스 토큰 재발급
+    /**
+     * 액세스 토큰 재발급
+     */
     @Transactional
     public ResponseEntity<?> reissueAccessToken(String refreshToken) {
 
@@ -98,13 +101,6 @@ public class OAuth2TokenService {
         User user = userRepository.findById(userId)
                 .orElseThrow(() -> new GlobalException(ErrorCode.USER_NOT_FOUND));
 
-        // 리프레시 토큰 검증
-        if (refreshToken == null || !jwtTokenProvider.validateToken(refreshToken)) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
-                    .body(ApiResponse
-                            .error("리프레시 토큰이 유효하지 않습니다.",
-                                    "UNAUTHORIZED"));
-        }
 
         // 새로운 액세스 토큰 발급
         String newAccessToken = jwtTokenProvider.createAccessToken(user);
@@ -114,7 +110,9 @@ public class OAuth2TokenService {
                 .body("새로운 액세스 토큰이 발급되었습니다.");
     }
 
-    // 리프레시 토큰 삭제
+    /**
+     * 리프레시 토큰 삭제
+     */
     public void deleteRefreshToken(Long userId) {
         redisTemplate.delete("REFRESH_TOKEN:" + userId);
     }
